introduce the possibility to have z0_ as a float
git-svn-id: https://svn.opendtect.org/od/trunk@38519 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/AttributeEngine/attribdatacubes.h
@@ -85,7 +85,7 @@ public:
 
     SamplingData<int>		inlsampling_;
     SamplingData<int>		crlsampling_;
-    int				z0_;
+    float			z0_;
     double			zstep_;
     TrcKeyZSampling		cubeSampling() const;
 				/*!<For convenience. The samling is changed
